Add a work-around for a race condition in Mono.
Patch from
https://github.com/Unity-Technologies/mono/commit/7e127d432c2ab2f936ff41b6cde714e34ad5fd23

The short version of this story is that Mono was using CREATE_SUSPENDED to do
synchronization, and this won't work anymore because some other part of the
code needs to do work on the new threads before they are unsuspended. So now
they are calling SuspendThread(GetCurrentThread()) from the new thread, and
we have a race where the thread can be resumed before it is suspended.

This makes problems less likely, but I think they can still happen if we
have a GC pause at a bad time. This is a serious problem on Windows, and Mono
developers are aware of it, so I'm hoping it'll be fixed by 2.12.0.