Remove MonoGame from the distribution.
While a free and binary-compatible XNA replacement would be really cool, it
should be done in a more general way than this (one which allows for use
outside of Wine). Mono should instead be fixed to work with the original
implementation of XNA, which is what we end up using in practice anyway.