Makefile: Expicitly disable vsscanf when compiling libSDL2.
msvcrt.dll doesn't support vsscanf. mingw-w64 uses a wrapper to convert it to
sscanf, but this wrapper is imperfect; it currently suffers from a bug
causing register corruption, and even with that fixed it copies a fixed
number of arguments to the stack, which will obviously fail if too many
arguments are provided and may cause stack underflows if too few are.

Instead just report to SDL that vsscanf is not supported, which is largely
true, and let it use its own implementation.