winforms update: Revert upstream merges since 4.9.4.
Some layout issues were reveled by the update to 5.0.0. While this
is probably fixable, it's not clear that what we brought in from
upstream has any value. It brings us further from what shipped in
.NET. If we need specific bugfixes from upstream, we can easily
bring them in, and I can keep watching upstream activity for
such things.

A few patches that may be relevant to us have been cherry-picked.

The Microsoft.VisualBasic.Forms assembly is not used, but I
thought it'd be useful to have in case we need to import from it.