monolite: Fixup some mscorlib refernces to System.
Mono moved the generic Stack and Queue types from System.dll to
mscorlib.dll. In order to move them back and have the bootstrap
process still work, we need monolite binaries that reference them
in System.dll, because we can't reference System.dll in mscorlib.dll.

This should just be a temporary hack.