Update Wine test results.
This includes a change to mark PingTest as unreliable. I found it
has started failing, and as it also fails in Wine stable, I
cannot attribute this to a change in Wine or Wine Mono.

I'm not sure what in the environment could have changed either,
but it seems clear that allowing CAP_NET_RAW for normal users is
not a typical configuration on Linux, so the expected results
should account for that.