Add option to build system for preferencing DWARF symbols.
This commit allows the build system to output DWARF symbols when
PREFER_DWARF_SYMBOLS=1 is set, for example, in user-config.make.

Works best with GCC MinGW rather than LLVM, so it is recommended to
also set MINGW_PATH (for example: MINGW_PATH=/usr/bin).

It also adds a LLVM_MINGW_PATH, which should be modified if a custom
LLVM path is to be used. MINGW_PATH now defaults to this value.

wpfgfx-netcore and some tests only compile with LLVM and always use
LLVM_MINGW_PATH. The rest will use MINGW_PATH which can be LLVM or GCC.